устранение дублирования кода во вью-моделях
