+++ app/src/main/java/dyachenko/kotlinmaterialdesign03/view/earth/EarthFragment.kt
@@ -12,6 +12,7 @@ import com.squareup.picasso.Callback
 import com.squareup.picasso.Picasso
 import dyachenko.kotlinmaterialdesign03.R
 import dyachenko.kotlinmaterialdesign03.databinding.EarthFragmentBinding
+import dyachenko.kotlinmaterialdesign03.model.earth.EarthResponseData
 import dyachenko.kotlinmaterialdesign03.model.settings.SettingsData
 import dyachenko.kotlinmaterialdesign03.util.*
 import dyachenko.kotlinmaterialdesign03.viewmodel.AppState
@@ -91,8 +92,8 @@ class EarthFragment : Fragment() {
 
     private fun renderData(data: AppState) = with(binding) {
         when (data) {
-            is AppState.SuccessEarth -> {
-                val responseData = data.responseData
+            is AppState.Success<*> -> {
+                val responseData = data.responseData as EarthResponseData
                 val url = responseData.url
 
                 Picasso
@@ -120,10 +121,6 @@ class EarthFragment : Fragment() {
                 earthRootView.showSnackBar(data.error.message ?: getString(R.string.error_msg),
                     getString(R.string.reload_msg),
                     { getData() })
-            }
-            is AppState.SuccessPOD -> {
-            }
-            is AppState.SuccessMars -> {
             }
         }
     }
