+++ app/src/main/java/dyachenko/kotlinmaterialdesign03/view/pod/PODFragment.kt
@@ -15,6 +15,7 @@ import com.squareup.picasso.Picasso
 import dyachenko.kotlinmaterialdesign03.R
 import dyachenko.kotlinmaterialdesign03.databinding.PodBottomSheetLayoutBinding
 import dyachenko.kotlinmaterialdesign03.databinding.PodFragmentBinding
+import dyachenko.kotlinmaterialdesign03.model.pod.PODResponseData
 import dyachenko.kotlinmaterialdesign03.model.settings.SettingsData
 import dyachenko.kotlinmaterialdesign03.util.*
 import dyachenko.kotlinmaterialdesign03.viewmodel.AppState
@@ -90,8 +91,8 @@ class PODFragment : Fragment() {
 
     private fun renderData(data: AppState) = with(binding) {
         when (data) {
-            is AppState.SuccessPOD -> {
-                val responseData = data.responseData
+            is AppState.Success<*> -> {
+                val responseData = data.responseData as PODResponseData
                 val url = responseData.url
 
                 if (responseData.mediaType == "image") {
@@ -132,10 +133,6 @@ class PODFragment : Fragment() {
                 podRootView.showSnackBar(data.error.message ?: getString(R.string.error_msg),
                     getString(R.string.reload_msg),
                     { getData() })
-            }
-            is AppState.SuccessEarth -> {
-            }
-            is AppState.SuccessMars -> {
             }
         }
     }
